requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.9
botocore==1.36.9
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.7
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wheel==0.45.1
